+++ UsuariosConexion/Daos/Usuario.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace UsuariosConexion
+namespace Daos
 {
     public class Usuario
     {
